+++ certificates/tests/test_views.py
@@ -19,8 +19,8 @@ from common.tests.util import raises_if
 from common.tests.util import view_is_subclass
 from common.tests.util import view_urlpattern_ids
 from common.validators import UpdateType
-from common.views import TamatoListView
-from common.views import TrackedModelDetailMixin
+from common.views.base import TamatoListView
+from common.views.mixins import TrackedModelDetailMixin
 
 pytestmark = pytest.mark.django_db
