+++ commodities/tests/test_views.py
@@ -20,7 +20,7 @@ from common.tests.util import get_class_based_view_urls_matching_url
 from common.tests.util import view_is_subclass
 from common.tests.util import view_urlpattern_ids
 from common.validators import UpdateType
-from common.views import TrackedModelDetailMixin
+from common.views.mixins import TrackedModelDetailMixin
 
 pytestmark = pytest.mark.django_db
