+++ common/urls.py
@@ -9,18 +9,18 @@ from django.urls import include
 from django.urls import path
 from django.urls import register_converter
 
-from common import views
 from common.path_converters import NumericSIDConverter
+from common.views import pages
 from measures import views as measure_views
 
 register_converter(NumericSIDConverter, "sid")
 
 urlpatterns = [
-    path("", views.HomeView.as_view(), name="home"),
-    path("search/", views.SearchPageView.as_view(), name="search-page"),
-    path("resources/", views.ResourcesView.as_view(), name="resources"),
-    path("pingdom/ping.xml", views.HealthCheckView.as_view(), name="healthcheck"),
-    path("app-info", views.AppInfoView.as_view(), name="app-info"),
+    path("", pages.HomeView.as_view(), name="home"),
+    path("search/", pages.SearchPageView.as_view(), name="search-page"),
+    path("resources/", pages.ResourcesView.as_view(), name="resources"),
+    path("pingdom/ping.xml", pages.HealthCheckView.as_view(), name="healthcheck"),
+    path("app-info", pages.AppInfoView.as_view(), name="app-info"),
     path(
         "duties/",
         measure_views.DutySentenceReference.as_view(),
@@ -28,11 +28,11 @@ urlpatterns = [
     ),
     path(
         "accessibility-statement",
-        views.AccessibilityStatementView.as_view(),
+        pages.AccessibilityStatementView.as_view(),
         name="accessibility-statement",
     ),
-    path("login", views.LoginView.as_view(), name="login"),
-    path("logout", views.LogoutView.as_view(), name="logout"),
+    path("login", pages.LoginView.as_view(), name="login"),
+    path("logout", pages.LogoutView.as_view(), name="logout"),
     path("api-auth/", include("rest_framework.urls")),
-    path("maintenance/", views.MaintenanceView.as_view(), name="maintenance"),
+    path("maintenance/", pages.MaintenanceView.as_view(), name="maintenance"),
 ]
