+++ geo_areas/views.py
@@ -12,12 +12,12 @@ from common.models.trackedmodel import TrackedModel
 from common.serializers import AutoCompleteSerializer
 from common.util import TaricDateRange
 from common.validators import UpdateType
-from common.views import DescriptionDeleteMixin
-from common.views import SortingMixin
-from common.views import TamatoListView
-from common.views import TrackedModelDetailMixin
-from common.views import TrackedModelDetailView
-from common.views import WithPaginationListMixin
+from common.views.base import TamatoListView
+from common.views.base import TrackedModelDetailView
+from common.views.mixins import DescriptionDeleteMixin
+from common.views.mixins import SortingMixin
+from common.views.mixins import TrackedModelDetailMixin
+from common.views.mixins import WithPaginationListMixin
 from geo_areas import business_rules
 from geo_areas import forms
 from geo_areas.filters import GeographicalAreaFilter
