+++ regulations/tests/test_views.py
@@ -16,8 +16,8 @@ from common.tests.util import valid_between_start_delta
 from common.tests.util import view_is_subclass
 from common.tests.util import view_urlpattern_ids
 from common.validators import UpdateType
-from common.views import TamatoListView
-from common.views import TrackedModelDetailMixin
+from common.views.base import TamatoListView
+from common.views.mixins import TrackedModelDetailMixin
 from regulations.models import Regulation
 from regulations.validators import RegulationUsage
 from regulations.views import RegulationDetailMeasures
